[NS-87] Move C++ code, pt 3
- Move `nullable_*_vector` and `handle_transfer` code to `core/`
