+++ src/main/resources/io/sparkcyclone/cpp/cyclone/core/nullable_varchar_vector.cc
@@ -17,16 +17,16 @@
  * limitations under the License.
  *
  */
-#include "cyclone/transfer-definitions.hpp"
 #include "cyclone/algorithm/bitset.hpp"
 #include "cyclone/algorithm/grouping.hpp"
+#include "cyclone/core/transfer-definitions.hpp"
 #include "frovedis/core/utility.hpp"
 #include "frovedis/text/char_int_conv.hpp"
 #include "frovedis/text/datetime_utility.hpp"
 #include "frovedis/text/dict.hpp"
 #include <stdlib.h>
-#include <iostream>
 #include <cstring>
+#include <iostream>
 
 nullable_varchar_vector * nullable_varchar_vector::allocate() {
   // Allocate
