+++ src/main/resources/io/sparkcyclone/cpp/cyclone/util/example.cc
@@ -17,15 +17,10 @@
  * limitations under the License.
  *
  */
-#include "cyclone/example.hpp"
-#include "tests/doctest.h"
+#include "cyclone/util/example.hpp"
 
-namespace cyclone::tests {
-  TEST_CASE("Testing the factorial function") {
-    CHECK(factorial(0) == 1);
-    CHECK(factorial(1) == 1);
-    CHECK(factorial(2) == 2);
-    CHECK(factorial(3) == 6);
-    CHECK(factorial(10) == 3628800);
+namespace cyclone::example {
+  int factorial(int number) {
+    return number > 1 ? factorial(number - 1) * number : 1;
   }
 }
