+++ src/main/resources/io/sparkcyclone/cpp/cyclone/util/example.hpp
@@ -17,10 +17,8 @@
  * limitations under the License.
  *
  */
-#include "cyclone/example.hpp"
+#pragma once
 
-namespace cyclone {
-  int factorial(int number) {
-    return number > 1 ? factorial(number - 1) * number : 1;
-  }
+namespace cyclone::example {
+  int factorial(int number);
 }
