+++ src/main/resources/io/sparkcyclone/cpp/cyclone/util/shm.hpp
@@ -19,6 +19,11 @@
  */
 #pragma once
 
-namespace cyclone {
-  int factorial(int number);
-}
+/*
+#ifdef __ve__
+
+extern "C" int attach_vh_shm(char *path, int32_t id, size_t size_mb, void **out_p, uint64_t *out_data_vehva);
+extern "C" int dettach_vh_shm(void *p, uint64_t data_vehva);
+
+#endif
+*/
