+++ src/main/resources/io/sparkcyclone/cpp/tests/core/nullable_scalar_vector_spec.cc
