+++ src/main/resources/io/sparkcyclone/cpp/tests/core/nullable_varchar_vector_spec.cc
